+++ src/main/java/wooteco/subway/exception/NotFoundStationException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception;
+
+import org.springframework.http.HttpStatus;
+
+public final class NotFoundStationException extends SubwayException{
+
+    public NotFoundStationException(String message) {
+        super(message, HttpStatus.NOT_FOUND);
+    }
+}
